+++ elements/stacinfo/stories/index.js
@@ -1,0 +1,8 @@
+export { default as PrimaryStory } from "./primary";
+export { default as SinglePropertyStory } from "./single-property";
+export { default as HeaderPropertiesStory } from "./header";
+export { default as TagsStory } from "./tags";
+export { default as FeaturedPropertiesStory } from "./featured-properties";
+export { default as FooterStory } from "./footer";
+export { default as CustomSlotContentStory } from "./custom-slot-content";
+export { default as UnstyledStory } from "./unstyled";