+++ elements/layercontrol/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [0.20.0](https://github.com/EOX-A/EOxElements/compare/layercontrol-v0.19.0...layercontrol-v0.20.0) (2024-08-12)
+
+
+### Features
+
+* Introduce Layercontrol datetime tool ([#1162](https://github.com/EOX-A/EOxElements/issues/1162)) ([6df8daa](https://github.com/EOX-A/EOxElements/commit/6df8daac8d8be6bf7d0a3447c940e2810734f739))
+
+## [0.19.0](https://github.com/EOX-A/EOxElements/compare/layercontrol-v0.18.1...layercontrol-v0.19.0) (2024-08-06)
+
+
+### Features
+
+* Allow passing `eox-map` DOM element to `for` property reactively ([#1152](https://github.com/EOX-A/EOxElements/issues/1152)) ([379d885](https://github.com/EOX-A/EOxElements/commit/379d885ddf14980e6b861172fbd066df36bf152d))
+
 ## [0.18.1](https://github.com/EOX-A/EOxElements/compare/layercontrol-v0.18.0...layercontrol-v0.18.1) (2024-07-31)