+++ elements/timecontrol/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [0.7.2](https://github.com/EOX-A/EOxElements/compare/timecontrol-v0.7.1...timecontrol-v0.7.2) (2024-08-19)
+
+
+### Bug Fixes
+
+* Render high-density slider ticks correctly ([#1221](https://github.com/EOX-A/EOxElements/issues/1221)) ([fc272f1](https://github.com/EOX-A/EOxElements/commit/fc272f11911faa7f574936b2b953252069f3e48d))
+
+## [0.7.1](https://github.com/EOX-A/EOxElements/compare/timecontrol-v0.7.0...timecontrol-v0.7.1) (2024-08-12)
+
+
+### Bug Fixes
+
+* Pre-calculate slider ticks to reduce loading time ([#1208](https://github.com/EOX-A/EOxElements/issues/1208)) ([a1a2fcb](https://github.com/EOX-A/EOxElements/commit/a1a2fcb4803e780d0e9fdc00c0dc4e43eb9fafed))
+
+## [0.7.0](https://github.com/EOX-A/EOxElements/compare/timecontrol-v0.6.1...timecontrol-v0.7.0) (2024-08-06)
+
+
+### Features
+
+* Allow passing `eox-map` DOM element to `for` property reactively ([#1152](https://github.com/EOX-A/EOxElements/issues/1152)) ([379d885](https://github.com/EOX-A/EOxElements/commit/379d885ddf14980e6b861172fbd066df36bf152d))
+
 ## [0.6.1](https://github.com/EOX-A/EOxElements/compare/timecontrol-v0.6.0...timecontrol-v0.6.1) (2024-07-30)